refactor: change location exceptions
