+++ Source/Application/Do-Svyazi.Message.Application.Abstractions/Exceptions/UnauthorizedException.cs
@@ -1,4 +1,4 @@
-﻿namespace Do_Svyazi.Message.Application.Abstractions.Tools;
+﻿namespace Do_Svyazi.Message.Application.Abstractions.Exceptions;
 
 public abstract class UnauthorizedException : ApplicationException
 {
